requirements.txt
pytest:
requests # Библиотека для HTTP-запросов

pytest # Фреймворк для тестирования
selenium # Инструмент для автоматизации браузера
webdriver-manager # Утилита для управления веб-драйверами
allure-pytest # Плагин для интеграции Allure с pytest
faker # Библиотека для генерации фейковых данных

flake8 # Линтер для проверки стиля кода